tracerca_query/python_grpc/requirements.txt
protobuf==3.20.3
grpcio
opentelemetry-proto
